refactor(AudioContext): move files around
