+++ SoundWave/Classes/AVAudioFileExtensions.swift
@@ -34,45 +34,3 @@ extension AVAudioFile {
 		return result
 	}
 }
-
-final class AudioContext {
-	public let audioURL: URL
-	public let totalSamples: Int
-	public let asset: AVAsset
-	public let assetTrack: AVAssetTrack
-
-	private init(audioURL: URL, totalSamples: Int, asset: AVAsset, assetTrack: AVAssetTrack) {
-		self.audioURL = audioURL
-		self.totalSamples = totalSamples
-		self.asset = asset
-		self.assetTrack = assetTrack
-	}
-
-	public static func load(fromAudioURL audioURL: URL, completionHandler: @escaping (_ audioContext: AudioContext?) -> ()) {
-		let asset = AVURLAsset(url: audioURL, options: [AVURLAssetPreferPreciseDurationAndTimingKey: NSNumber(value: true as Bool)])
-
-		guard let assetTrack = asset.tracks(withMediaType: AVMediaType.audio).first else {
-			fatalError("Couldn't load AVAssetTrack")
-		}
-
-		asset.loadValuesAsynchronously(forKeys: ["duration"]) {
-			var error: NSError?
-			let status = asset.statusOfValue(forKey: "duration", error: &error)
-			switch status {
-			case .loaded:
-				guard let formatDescriptions = assetTrack.formatDescriptions as? [CMAudioFormatDescription],
-					let audioFormatDesc = formatDescriptions.first,
-					let asbd = CMAudioFormatDescriptionGetStreamBasicDescription(audioFormatDesc) else
-				{
-					break
-				}
-				let totalSamples = Int((asbd.pointee.mSampleRate) * Float64(asset.duration.value) / Float64(asset.duration.timescale))
-				let audioContext = AudioContext(audioURL: audioURL, totalSamples: totalSamples, asset: asset, assetTrack: assetTrack)
-				completionHandler(audioContext)
-			case .failed, .cancelled, .loading, .unknown:
-				print("Couldn't load asset: \(error?.localizedDescription ?? "Unknown error")")
-				completionHandler(nil)
-			}
-		}
-	}
-}
